+++ OneBlink.SDK/models/FormApprovalCannedResponse.cs
@@ -1,0 +1,18 @@
+namespace OneBlink.SDK.Model
+{
+    public class FormApprovalCannedResponse
+    {
+        public string key
+        {
+            get; set;
+        }
+        public string label
+        {
+            get; set;
+        }
+        public string notes
+        {
+            get; set;
+        }
+    }
+}