+++ OneBlink.SDK/models/FormSubmissionApproval.cs
@@ -60,5 +60,9 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public string cannedResponseKey
+        {
+            get; set;
+        }
     }
 }